feat: Init almost complete resampling code.
I have to still create/overwrite the list of particles with the
resampled particles. I cannot simply use numpy indexing because while
that while create a copy of the original list, it will not create a copy
of each particle, which is what I want to do.

For the above I need to either:
* Add a set_exact_loc() to Robot, or better
* Derive Particle from Robot and add set_exact_loc() to it.