fix: Add bookshelf to hopefully improve AMCL.
I was noticing that the robot would sometimes substantially deviate from
the planned path. Investigating in the kb led me to this question -
https://knowledge.udacity.com/questions/63382.
From the answer it seems that the above behavior can be caused because
of poor localization, and a way to mitigate this poorness could be to
add more "landmarks" to the environment that can help the robot
differentiate between neighborhoods. It's a bit hand-wavy but adding an
extra bookshelf does seem to help.